# [crypto/ed25519] Preallocate batch size for verification